code/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
httpx>=0.25.0
numpy>=1.22.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.0.0
joblib>=1.1.0
python-multipart>=0.0.5
email-validator>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
jwt>=1.3.1
python-jose>=3.3.0
passlib>=1.7.4